Added a test opmode
I added a test opmode: Team2Test.java, which includes functions for moving, strafing, and turning, but doesn't call any functions. It is for testing different parts of our drivetrain individually. Controller support needs to be added, if not already, and some of the strafing functions need to be tested. The functions are suitable, after testing, for the simple game that we are competing with Team 3 after we add all the hardware components we need to push a cone and controller support